Adição de novos agendamentos com relacionamento conferido.
