+++ src/main/java/com/api/agendamedicaback/resources/UsuarioResource.java
@@ -3,7 +3,7 @@ package com.api.agendamedicaback.resources;
 
 import com.api.agendamedicaback.domain.Usuario;
 import com.api.agendamedicaback.domain.dtos.UsuarioDTO;
-import com.api.agendamedicaback.resources.services.UsuarioService;
+import com.api.agendamedicaback.services.UsuarioService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
