+++ src/main/java/com/api/agendamedicaback/services/exceptions/DataIntegrityViolationException.java
@@ -1,4 +1,4 @@
-package com.api.agendamedicaback.resources.services.exceptions;
+package com.api.agendamedicaback.services.exceptions;
 
 public class DataIntegrityViolationException extends RuntimeException{
     private static final long serialVersionUID = 1L;
