+++ src/main/java/com/api/agendamedicaback/services/exceptions/ObjectNotFoundException.java
@@ -1,4 +1,4 @@
-package com.api.agendamedicaback.resources.services.exceptions;
+package com.api.agendamedicaback.services.exceptions;
 
 public class ObjectNotFoundException extends RuntimeException {
     private static final long serialVersionUID = 1L;
